+++ mods/BrewLAN/documentation/Wiki Data.lua
@@ -374,6 +374,9 @@ UnitData = {
         Trivia = '"Simurgh" was a commissioned unit based on the Arsenal Bird from *Ace Combat 7*.',
     },
     SRA4212 = {
+        Videos = {
+            {YouTube = 'G-dNAHv-jA4', 'Time-lapse'},
+        },
         Trivia = {
             '"Orlok" was a commissioned unit.',
             'The name is taken from the name of the count from the 1922 film *Nosferatu: A Symphony of Horror*; Count Orlok.',